requirements.txt
-i https://pypi.org/simple
async-generator==1.10 ; python_version >= '3.5'
attrs==23.1.0 ; python_version >= '3.7'
authlib==1.2.0
blessed==1.20.0 ; python_version >= '2.7'
certifi==2023.5.7 ; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
colorama==0.4.6
cryptography==41.0.1 ; python_version >= '3.7'
dateparser==1.1.8
exceptiongroup==1.1.1 ; python_version >= '3.7'
h11==0.14.0 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
inquirer==3.1.3
markdown-it-py==3.0.0 ; python_version >= '3.8'
mdurl==0.1.2 ; python_version >= '3.7'
outcome==1.2.0 ; python_version >= '3.7'
pycparser==2.21
pygments==2.15.1 ; python_version >= '3.7'
pysocks==1.7.1
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
python-editor==1.0.4
pytz==2023.3
pyyaml==6.0
readchar==4.0.5 ; python_version >= '3.7'
regex==2023.6.3 ; python_version >= '3.6'
requests==2.31.0
rich==13.4.2
selenium==4.10.0
selenium-stealth==1.0.6
setuptools==67.8.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
sortedcontainers==2.4.0
trio==0.22.0 ; python_version >= '3.7'
trio-websocket==0.10.3 ; python_version >= '3.7'
tzlocal==5.0.1 ; python_version >= '3.7'
urllib3==2.0.3 ; python_version >= '3.7'
wcwidth==0.2.6
wsproto==1.2.0 ; python_full_version >= '3.7.0'
